fix cbvb
